requirements.txt
# > pip list
# > pip list | findstr pandas
streamlit==1.55.0
pandas==2.3.3
sqlalchemy==2.0.48
langchain==1.2.10
langchain-groq==1.1.2
langchain-community==0.4.1
python-dotenv==1.2.2
matplotlib==3.10.8
ipykernel==7.2.0
plotly==6.6.0